fix(purchase): revert qty->weight conflation on Goods Receipt, add real weight tracking
Receive Now/Qty is a unit count (zak/lot/unit), not a truck-scale weight -
the prior sprint (f294b46) widened qty to decimal(15,2) across the whole
Purchase chain on that false premise. Reverts qty back to integer end to
end (PO/GR/Invoice/Return qty, Item.current_stock, StockLedger), then adds
actual_weight/weight_unit/weighbridge_ref as a true optional sibling field
on Goods Receipt lines that never touches qty/rate/amount/stock math.

Also fixes two structural gaps this exposed: a from-PO Goods Receipt line
can now be added/removed/edited (previously fixed 1:1 to PO lines), and
the same PO item can appear on more than one line (different truck loads) -
combined qty per item is validated against outstanding, not each line
alone. Item dropdown in from-PO mode is restricted to that PO's own lines.

Adds GoodsReceiptExport + a weight column on the Goods Receipt report,
mirroring the existing Purchase Invoice export pattern.

+++ backend/laravel-api/app/Repositories/StockLedgerRepository.php
@@ -13,9 +13,9 @@ class StockLedgerRepository extends BaseRepository
         parent::__construct($model);
     }
 
-    public function latestBalance(string $itemId, string $warehouseId): float
+    public function latestBalance(string $itemId, string $warehouseId): int
     {
-        return (float) $this->model->query()
+        return (int) $this->model->query()
             ->where('item_id', $itemId)
             ->where('warehouse_id', $warehouseId)
             ->orderByDesc('posting_datetime')
@@ -57,14 +57,14 @@ class StockLedgerRepository extends BaseRepository
      * across every warehouse, not just the one most recently touched —
      * sum each warehouse's latest balance, not just the latest write.
      */
-    public function totalBalanceForItem(string $itemId): float
+    public function totalBalanceForItem(string $itemId): int
     {
         $warehouseIds = $this->model->query()
             ->where('item_id', $itemId)
             ->distinct()
             ->pluck('warehouse_id');
 
-        $total = 0.0;
+        $total = 0;
 
         foreach ($warehouseIds as $warehouseId) {
             $total += $this->latestBalance($itemId, $warehouseId);
@@ -79,9 +79,9 @@ class StockLedgerRepository extends BaseRepository
      * currently say") where taking FOR UPDATE would hold a lock for no
      * real reason and risks contention between unrelated concurrent drafts.
      */
-    public function latestBalanceUnlocked(string $itemId, string $warehouseId): float
+    public function latestBalanceUnlocked(string $itemId, string $warehouseId): int
     {
-        return (float) $this->model->query()
+        return (int) $this->model->query()
             ->where('item_id', $itemId)
             ->where('warehouse_id', $warehouseId)
             ->orderByDesc('posting_datetime')
